+++ package.json
@@ -46,6 +46,7 @@
     "react-native-wkwebview-reborn": "^1.20.0",
     "react-navigation": "^2.0.1",
     "recompose": "^0.27.0",
+    "subscriptions-transport-ws": "^0.9.10",
     "url-parse": "^1.4.0"
   },
   "devDependencies": {
